requirements.txt
numpy==1.20.1
tensorflow-gpu==2.2.0
Keras==2.4.3
jieba==0.42.1
matplotlib==3.4.1
pandas==1.2.2
scikit-learn==0.24.1
stylecloud==0.5.2